feat!: upgrade next, react and react-dom, move react/next dependency checker from payload to next package
