+++ packages/next/package.json
@@ -88,23 +88,23 @@
     "@babel/preset-env": "^7.24.5",
     "@babel/preset-react": "^7.24.1",
     "@babel/preset-typescript": "^7.24.1",
-    "@next/eslint-plugin-next": "15.0.0-canary.104",
+    "@next/eslint-plugin-next": "15.0.0-canary.160",
     "@payloadcms/eslint-config": "workspace:*",
     "@types/busboy": "1.5.4",
     "@types/react": "npm:types-react@19.0.0-rc.0",
     "@types/react-dom": "npm:types-react-dom@19.0.0-rc.0",
     "@types/uuid": "10.0.0",
     "@types/ws": "^8.5.10",
-    "babel-plugin-react-compiler": "0.0.0-experimental-48eb8f4-20240822",
+    "babel-plugin-react-compiler": "0.0.0-experimental-24ec0eb-20240918",
     "esbuild": "0.23.1",
     "esbuild-sass-plugin": "3.3.1",
-    "eslint-plugin-react-compiler": "0.0.0-experimental-72f06b2-20240822",
+    "eslint-plugin-react-compiler": "0.0.0-experimental-7670337-20240918",
     "payload": "workspace:*",
     "swc-plugin-transform-remove-imports": "1.15.0"
   },
   "peerDependencies": {
     "graphql": "^16.8.1",
-    "next": "^15.0.0-canary.104",
+    "next": "^15.0.0-canary.160",
     "payload": "workspace:*"
   },
   "engines": {
