+++ packages/richtext-lexical/package.json
@@ -88,7 +88,7 @@
     "babel-plugin-transform-remove-imports": "^1.8.0",
     "esbuild": "0.23.1",
     "esbuild-sass-plugin": "3.3.1",
-    "eslint-plugin-react-compiler": "0.0.0-experimental-72f06b2-20240822",
+    "eslint-plugin-react-compiler": "0.0.0-experimental-7670337-20240918",
     "payload": "workspace:*",
     "swc-plugin-transform-remove-imports": "1.15.0"
   },
@@ -108,8 +108,8 @@
     "@payloadcms/next": "workspace:*",
     "lexical": "0.17.0",
     "payload": "workspace:*",
-    "react": "^19.0.0 || ^19.0.0-rc-06d0b89e-20240801",
-    "react-dom": "^19.0.0 || ^19.0.0-rc-06d0b89e-20240801"
+    "react": "^19.0.0 || ^19.0.0-rc-5dcb0097-20240918",
+    "react-dom": "^19.0.0 || ^19.0.0-rc-5dcb0097-20240918"
   },
   "engines": {
     "node": "^18.20.2 || >=20.9.0"
