+++ packages/ui/package.json
@@ -109,17 +109,17 @@
     "@types/react-datepicker": "6.2.0",
     "@types/react-dom": "npm:types-react-dom@19.0.0-rc.0",
     "@types/uuid": "10.0.0",
-    "babel-plugin-react-compiler": "0.0.0-experimental-48eb8f4-20240822",
+    "babel-plugin-react-compiler": "0.0.0-experimental-24ec0eb-20240918",
     "esbuild": "0.23.1",
     "esbuild-sass-plugin": "3.3.1",
-    "eslint-plugin-react-compiler": "0.0.0-experimental-72f06b2-20240822",
+    "eslint-plugin-react-compiler": "0.0.0-experimental-7670337-20240918",
     "payload": "workspace:*"
   },
   "peerDependencies": {
-    "next": "^15.0.0-canary.104",
+    "next": "^15.0.0-canary.160",
     "payload": "workspace:*",
-    "react": "^19.0.0 || ^19.0.0-rc-06d0b89e-20240801",
-    "react-dom": "^19.0.0 || ^19.0.0-rc-06d0b89e-20240801"
+    "react": "^19.0.0 || ^19.0.0-rc-5dcb0097-20240918",
+    "react-dom": "^19.0.0 || ^19.0.0-rc-5dcb0097-20240918"
   },
   "engines": {
     "node": "^18.20.2 || >=20.9.0"
