+++ templates/website/package.json
@@ -40,12 +40,12 @@
     "jsonwebtoken": "9.0.2",
     "lexical": "0.17.0",
     "lucide-react": "^0.378.0",
-    "next": "15.0.0-canary.104",
+    "next": "15.0.0-canary.160",
     "payload": "beta",
     "payload-admin-bar": "^1.0.6",
     "prism-react-renderer": "^2.3.1",
-    "react": "19.0.0-rc-06d0b89e-20240801",
-    "react-dom": "19.0.0-rc-06d0b89e-20240801",
+    "react": "19.0.0-rc-5dcb0097-20240918",
+    "react-dom": "19.0.0-rc-5dcb0097-20240918",
     "react-hook-form": "7.45.4",
     "sharp": "0.32.6",
     "tailwind-merge": "^2.3.0",
@@ -63,7 +63,7 @@
     "autoprefixer": "^10.4.19",
     "copyfiles": "^2.4.1",
     "eslint": "^8",
-    "eslint-config-next": "15.0.0-canary.104",
+    "eslint-config-next": "15.0.0-canary.160",
     "postcss": "^8.4.38",
     "prettier": "^3.0.3",
     "tailwindcss": "^3.4.3",
