+++ templates/with-payload-cloud/package.json
@@ -21,10 +21,10 @@
     "@payloadcms/richtext-lexical": "beta",
     "cross-env": "^7.0.3",
     "graphql": "^16.8.1",
-    "next": "15.0.0-canary.104",
+    "next": "15.0.0-canary.160",
     "payload": "beta",
-    "react": "19.0.0-rc-06d0b89e-20240801",
-    "react-dom": "19.0.0-rc-06d0b89e-20240801",
+    "react": "19.0.0-rc-5dcb0097-20240918",
+    "react-dom": "19.0.0-rc-5dcb0097-20240918",
     "sharp": "0.32.6"
   },
   "devDependencies": {
@@ -32,7 +32,7 @@
     "@types/react": "npm:types-react@19.0.0-rc.0",
     "@types/react-dom": "npm:types-react-dom@19.0.0-rc.0",
     "eslint": "^8",
-    "eslint-config-next": "15.0.0-canary.104",
+    "eslint-config-next": "15.0.0-canary.160",
     "typescript": "5.6.2"
   },
   "engines": {
