+++ test/package.json
@@ -60,7 +60,7 @@
     "@sentry/react": "^7.77.0",
     "@types/react": "npm:types-react@19.0.0-rc.0",
     "@types/react-dom": "npm:types-react-dom@19.0.0-rc.0",
-    "babel-plugin-react-compiler": "0.0.0-experimental-48eb8f4-20240822",
+    "babel-plugin-react-compiler": "0.0.0-experimental-24ec0eb-20240918",
     "comment-json": "^4.2.3",
     "create-payload-app": "workspace:*",
     "dotenv": "16.4.5",
@@ -71,11 +71,11 @@
     "http-status": "1.6.2",
     "jwt-decode": "4.0.0",
     "lexical": "0.17.0",
-    "next": "15.0.0-canary.104",
+    "next": "15.0.0-canary.160",
     "payload": "workspace:*",
     "qs-esm": "7.0.2",
-    "react": "19.0.0-rc-06d0b89e-20240801",
-    "react-dom": "19.0.0-rc-06d0b89e-20240801",
+    "react": "19.0.0-rc-5dcb0097-20240918",
+    "react-dom": "19.0.0-rc-5dcb0097-20240918",
     "server-only": "^0.0.1",
     "slate": "0.91.4",
     "tempy": "^1.0.1",
